Stop delete dubber if it is used by another part of the app

+++ src/app/components/dubber/dubbers-list/dubbers-list.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { DubberModel } from '../dubber-model';
 import { DubberService } from '../dubbers.service';
-import {NgForm} from '@angular/forms';
+import { FilmService } from '../../film/film.service';
 
 @Component({
   templateUrl: './dubbers-list.component.html',
@@ -10,14 +9,30 @@ import {NgForm} from '@angular/forms';
 
 export class DubbersListComponent implements OnInit {
 
-  constructor(private dubberService: DubberService) {}
+  constructor(
+    private dubberService: DubberService,
+    private filmService: FilmService
+  ) {}
 
   onDelete(dubber) {
-    this.dubberService.delete(dubber);
+    let filmsList = this.filmService.filmsList;
+    let dubberId = dubber.id.toString();
+    let filmDubbersID = [];
+    for (let film of filmsList) {
+      for (let dubber of film.dubbers) {
+        filmDubbersID.push(dubber.id);
+      }
+    };
+    if(filmDubbersID.includes(dubberId)) {
+      alert("You can't delete it, because this dubber is using!");
+    } else {
+      this.dubberService.delete(dubber);
+    }
   }
 
   ngOnInit() {
     this.dubberService.getAll();
+    this.filmService.getAll();
   }
 
 }
